airflow/requirements.txt
lxml
python-decouple
apache-airflow-providers-apache-spark==4.0.0
requests
pandas

